+++ api/src/main/java/org/allaymc/api/world/data/FireworkType.java
@@ -1,0 +1,14 @@
+package org.allaymc.api.world.data;
+
+/**
+ * FireworkShape represents the shape of a firework.
+ *
+ * @author daoge_cmd
+ */
+public enum FireworkType {
+    SMALL_SPHERE,
+    HUGE_SPHERE,
+    STAR,
+    CREEPER_HEAD,
+    BURST
+}