+++ server/src/main/java/org/allaymc/server/item/component/ItemWritableBookBaseComponentImpl.java
@@ -82,9 +82,8 @@ public class ItemWritableBookBaseComponentImpl extends ItemBaseComponentImpl imp
 
     @Override
     public NbtMap saveExtraTag() {
-        var extraTag = super.saveExtraTag();
-        var builder = extraTag != null ? extraTag.toBuilder() : NbtMap.builder();
-        return builder.putList(TAG_PAGES, NbtType.COMPOUND, this.pages.stream().map(page -> NbtMap.builder().putString(TAG_TEXT, page).build()).toList()).build();
+        return super.saveExtraTag().toBuilder()
+                .putList(TAG_PAGES, NbtType.COMPOUND, this.pages.stream().map(page -> NbtMap.builder().putString(TAG_TEXT, page).build()).toList()).build();
     }
 
     @Override